Clean up team name input
