+++ src/client/team/components/team-details.tsx
@@ -1,29 +1,20 @@
-import React, { useState } from 'react';
-
+import { Button } from '@material-ui/core';
+import React from 'react';
+import { useRecoilValue } from 'recoil';
+import { useUpdateTeam } from '../../api/hooks/use-update-team';
+import { golfersState } from '../../app';
 import { useManageGolfers } from '../hooks/use-manage-golfers';
-import { useRecoilState, useRecoilValue } from 'recoil';
-import { teamNameState, teamGolfersIdsState } from '../state/atoms';
+import { teamGolfersIdsState } from '../state/atoms';
+import { teamState } from '../state/selectors';
 import { TeamList } from './team-list';
 import { TeamName } from './team-name';
-import { EditName } from './edit-name';
-import { golfersState } from '../../app';
-import { Button } from '@material-ui/core';
-import { useUpdateTeam } from '../../api/hooks/use-update-team';
-import { teamState } from '../state/selectors';
 
 export const TeamDetails = () => {
   const allGolfers = useRecoilValue(golfersState);
   const selectedGolferIds = useRecoilValue(teamGolfersIdsState);
   const { removeGolfer } = useManageGolfers();
-  const [teamName, setTeamName] = useRecoilState(teamNameState);
-  const [editMode, setEditMode] = useState(false);
   const updateTeam = useUpdateTeam();
   const teamDetails = useRecoilValue(teamState);
-
-  const handleNameChange = (newName: string) => {
-    setTeamName(newName);
-    setEditMode(false);
-  };
 
   const onSave = () => {
     updateTeam(teamDetails);
@@ -31,11 +22,7 @@ export const TeamDetails = () => {
 
   return (
     <>
-      {editMode ? (
-        <EditName teamName={teamName} handleNameChange={handleNameChange} />
-      ) : (
-        <TeamName teamName={teamName} setEditMode={setEditMode} />
-      )}
+      <TeamName />
       <TeamList
         allGolfers={allGolfers}
         selectedGolferIds={selectedGolferIds}
